+++ src/testWithSpringBoot_2_5/kotlin/org/openrewrite/java/spring/boot2/InitilizationOrderingTest.kt
